backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.35.94
celery==5.4.0
redis==5.2.1
spotipy==2.25.0
pydantic-settings==2.7.1

